Fix logic of expired links
- Move the logic for checking expired links to MegaClient instead of
MegaApi.
- `MegaNode::getPublicLink()` fix: if method is `const`, polymorphism
is broken.
- New test to check regeneration of existing public links.

+++ examples/megacli.cpp
@@ -2957,7 +2957,7 @@ static void process_line(char* l)
                                     del = (words[2] == "del");
                                     if (!del)
                                     {
-                                        ets = atoi(words[2].c_str());
+                                        ets = atol(words[2].c_str());
                                     }
                                 }
